demo using a single notebook and simplified codebase

+++ code/train.py
@@ -6,11 +6,11 @@ from dataclasses import dataclass, field, asdict
 from typing import Optional, List
 
 import numpy as np
-
 import sys
 import os
-
+import pickle
 import logging
+
 logger = logging.getLogger(__name__)
 
 from transformers import (
@@ -21,8 +21,9 @@ from transformers import (
 )
 
 import transformers
-from torch.utils.data import DataLoader, WeightedRandomSampler
+from torch.utils.data import DataLoader, WeightedRandomSampler  
 from scipy.special import softmax
+from scipy.special import expit as sigmoid
 
 import datasets
 
@@ -42,23 +43,27 @@ class ModelArguments:
     experiment_name: str = field(
         metadata={"help": "The experiment name. Use as suffix for output paths"},
     )
+    
+    valid_labels: str = field(
+        metadata={"help": "One of: ['5-char', '4-char', '4-char-top50']"},
+    )
 
     model_dir: Optional[str] = field(
-        default='/content/gdrive/Shareddrives/PROJECT_ROOT_DIR/models',
+        default='/content/gdrive/Shareddrives/cs224n-final-project/models',
         metadata={"help": "Full path to /models directory"},
     )
 
-    cache_dir: str = field(
-        default='/content/tmp',
-        metadata={"help": (
-            "The directory to save checkpoints during training."
-            "We copy this direcotry to model_dir after the training"
-            "to avoid unnessary syncrnization with Google Drive for file during training."),
-        },
-    )  
+    # cache_dir: str = field(
+    #     default='/content/tmp',
+    #     metadata={"help": (
+    #         "The directory to save checkpoints during training."
+    #         "We copy this direcotry to model_dir after the training"
+    #         "to avoid unnessary syncrnization with Google Drive for file during training."),
+    #     },
+    # )  
 
     data_dir: Optional[str] = field(
-        default='/content/gdrive/Shareddrives/PROJECT_ROOT_DIR/injury-icd-dataset',
+        default='/content/gdrive/Shareddrives/cs224n-final-project/injury-icd-dataset',
         metadata={"help": "Full path to /injury-icd-dataset directory"},
     )
 
@@ -81,8 +86,8 @@ class TraumalIcdTrainingArguments:
     num_train_epochs: float = field(default=20.0, metadata={"help": "Total number of training epochs to perform."})
     warmup_steps: int = field(default=1000, metadata={"help": "Linear warmup over warmup_steps."})
     metric_for_best_model: str = field(default="eval_f1_score_weighted", metadata={"help": "The metric for selecting the best model for final evaluation."})
-    is_pretrain: bool = field(default=False, metadata={"help": "Whether to pretrain model on 5-char ICD10 codes."})
-    is_evaluate: bool = field(default=False, metadata={"help": "Whether just to evaluate a previously trained and saved model."})
+    train_on_full: bool = field(default=False, metadata={"help": "Whether train on both the train and validation splits, and eval on holdout test split."})
+    is_evaluate: bool = field(default=False, metadata={"help": "Whether to just evaluate a previously saved checkpoint (and not train anything)."})
 
     log_level: Optional[str] = field(
         default=logging.INFO,
@@ -103,8 +108,8 @@ class TraumalIcdTrainer(Trainer):
 
         training_args = TrainingArguments(
             output_dir=output_dir, 
-            per_device_train_batch_size=traumal_icd_training_args.per_device_train_batch_size,
-            per_device_eval_batch_size=traumal_icd_training_args.per_device_eval_batch_size,
+            per_device_train_batch_size=int(traumal_icd_training_args.per_device_train_batch_size),
+            per_device_eval_batch_size=int(traumal_icd_training_args.per_device_eval_batch_size),
             evaluation_strategy="epoch",
             metric_for_best_model=traumal_icd_training_args.metric_for_best_model,
             greater_is_better=True,
@@ -136,13 +141,14 @@ class TraumalIcdTrainer(Trainer):
         )
 
         self.icd_codes = icd_codes
+        self.output_dir = output_dir
 
     def evaluate(self, eval_dataset = None, ignore_keys = None, metric_key_prefix = "eval"):
         if eval_dataset is None:
             eval_dataset = self.eval_dataset
-
+        self.save_state()
         eval_dataloader = self.get_eval_dataloader(eval_dataset)
-        logger.info("Performing evaluation...")
+        logger.info("Saved trainer state and performing evaluation...")
         eval_loop = self.prediction_loop if self.args.use_legacy_prediction_loop else self.evaluation_loop
         output = eval_loop(
             eval_dataloader,
@@ -155,20 +161,38 @@ class TraumalIcdTrainer(Trainer):
         label_pos = dict((x,y) for y, x in enumerate(self.icd_codes.keys()))
         patient_pos = dict((x,y) for y, x in enumerate(list(set(eval_dataset['patient_id']))))
         predictions = softmax(output.predictions, axis=1)[:, 1]
+        # predictions = softmax(output.predictions, axis=1)
+        # predictions = sigmoid(output.predictions)
         targets = np.zeros((len(patient_pos), len(label_pos)))
         probs = np.zeros((len(patient_pos), len(label_pos)))
+        logger.info(f"Shape of predictions {predictions.shape}")
+        logger.info(f"Shape of targets {targets.shape}")
+        logger.info(f"Shape of probs {probs.shape}")
+        # for i, row in enumerate(eval_dataset):
+        #     assert row['label'] == output.label_ids[i]
+        #     targets[patient_pos[row['patient_id']], label_pos[row['icd_code']]] = row['label']
+        #     probs[patient_pos[row['patient_id']], label_pos[row['icd_code']]] = predictions[i]
 
         for i, row in enumerate(eval_dataset):
-            assert row['label'] == output.label_ids[i]
-            targets[patient_pos[row['patient_id']], label_pos[row['icd_code']]] = row['label']
-            probs[patient_pos[row['patient_id']], label_pos[row['icd_code']]] = predictions[i]
-
+        #   for j, icd_code in enumerate(row['icd_code']): 
+            icd_code = row['icd_code']
+            targets[patient_pos[row['patient_id']], label_pos[icd_code]] = row['label']
+            # probs[patient_pos[row['patient_id']], label_pos[icd_code]] = predictions[i, j]
+            probs[patient_pos[row['patient_id']], label_pos[icd_code]] = predictions[i]
+            # probs[patient_pos[row['patient_id']], label_pos[icd_code]] = predictions[i, 1]
         output.metrics.update(compute_metrics(targets, probs, prefix=metric_key_prefix))
         self.log(output.metrics)
 
         self.control = self.callback_handler.on_evaluate(self.args, self.state, self.control, output.metrics)
+        eval_pickle_path = os.path.join(self.output_dir, "eval_results.pickle")
+        eval_results = {"metrics": output.metrics, "probs": probs, "targets": targets, "patient_pos": patient_pos, "label_pos": label_pos}
+        with open(eval_pickle_path, 'wb') as f:
+            pickle.dump(eval_results, f)
+        logger.info(f"The evaluation results have been saved to to: {eval_pickle_path}")
 
         return output.metrics
+    
+
 
     def get_train_dataloader(self) -> DataLoader:
         train_ds = self.train_dataset
@@ -219,17 +243,17 @@ def main():
         level=logging.NOTSET
     )
     
-    assert not (training_args.is_evaluate and training_args.is_pretrain), "Please choose only one of is_pretrain or is_evaluate"
     ds_path = os.path.join(model_args.model_dir, model_args.model_name, model_args.experiment_name, 'encoded_ds')
-    ds, label_names = load_ds(model_args.data_dir, is_pretrain=training_args.is_pretrain)
+    ds, label_names = load_ds(model_args.data_dir, model_args.valid_labels, training_args.train_on_full)
+    logger.info(f"Loaded dataset with size {ds.shape}")
     
     logger.info(f'Loading base model and tokenizer: {model_args.model_name}...')
-    tokenizer = load_tokenizer(model_args.model_name, model_args.model_dir)
-    model = load_model(model_args.model_name, model_args.model_dir)
+    tokenizer = load_tokenizer(model_args.model_name, model_args.model_dir, model_args.experiment_name)
+    model = load_model(model_args.model_name, model_args.model_dir, model_args.experiment_name)
 
     if not os.path.exists(ds_path):
         logger.info(f'No tokenized dataset exist yet, so generating context-label pairs...')
-        pairs_ds = generate_pairs_ds(ds, label_names, model_args.data_dir, training_args.is_pretrain)
+        pairs_ds = generate_pairs_ds(ds, label_names, model_args.data_dir)
         logger.info(f'Finished generating context-label pairs, now tokenizing them into dataset...')
         ds_encoded = encode_ds(pairs_ds, tokenizer, model.config.max_position_embeddings)
         ds_encoded.save_to_disk(ds_path)
@@ -243,7 +267,7 @@ def main():
     logging_dir = os.path.join(model_args.model_dir, 'runs', model_args.model_name, 
         model_args.experiment_name)
 
-    output_dir = os.path.join(model_args.cache_dir, model_args.model_name, 
+    output_dir = os.path.join(model_args.model_dir, model_args.model_name, 
         model_args.experiment_name)
    
     output_model_dir = os.path.join(model_args.model_dir,
@@ -251,15 +275,13 @@ def main():
                                 model_args.experiment_name)
     
 
-    eval_ds_reduced = ds_encoded['reduced_validation']
-    eval_ds_full = ds_encoded['validation']
-
-    # patient_id = eval_ds[0]['patient_id']
-    # eval_ds = eval_ds.filter(lambda row: row['patient_id'] in ['1145', '1257'])
+    eval_ds_reduced = ds_encoded['reduced_eval']
+    eval_ds_full = ds_encoded['eval']
 
     trainer = TraumalIcdTrainer(model, tokenizer, output_dir, label_names,
                             train_ds=ds_encoded['train'],
                             eval_dataset=eval_ds_reduced,
+                            # eval_dataset=eval_ds_full,
                             logging_dir=logging_dir,
                             traumal_icd_training_args=training_args)
 
@@ -269,17 +291,19 @@ def main():
     
         trainer.save_state()
         save_args(output_dir, model_args, training_args, trainer.args)  
+        
+    if not training_args.is_evaluate:
+        logger.info(f'Saving...')
+        if os.path.exists(output_model_dir):
+            shutil.rmtree(output_model_dir)
+        shutil.copytree(output_dir, output_model_dir) 
     
-    logger.info('Evaluating on the full validation dataset...')
-    validation_metrics = trainer.evaluate(eval_ds_full)
+    logger.info('Evaluating on the full test/validation dataset...')
+    validation_metrics = trainer.evaluate(eval_ds_full) #, eval_args={"model_args": model_args, "training_args": training_args})
     logger.info(validation_metrics)
     with open(os.path.join(output_dir, 'best_scores.json'), 'w') as f:
         json.dump(validation_metrics, f)
-
-    logger.info(f'Saving...')
-    if os.path.exists(output_model_dir):
-        shutil.rmtree(output_model_dir)
-    shutil.copytree(output_dir, output_model_dir) 
+        
 
 if __name__ == "__main__":
     main()
